Fix CPU minimal training validation and document graph optimization tasks
This commit enables minimal CPU training for validation and identifies
performance issues with graph-based architectures for future work.

FIXES:
1. Fixed edge_index tensor type conversion for PyTorch Geometric
   - Added .long() conversion in configurable_extractor.py for all graph encoders
   - Fixed error: 'tensors used as indices must be long, int, byte or bool tensors'
   - Applied to HGTEncoder, SimplifiedHGTEncoder, GAT, and GCN

2. Dynamic n_steps/batch_size adjustment for minimal training
   - Modified architecture_trainer.py to handle very small total_timesteps
   - Automatically adjusts n_steps to max(total_timesteps // num_envs, 1)
   - Adjusts batch_size to match when needed for proper training
   - Enables CPU validation with --total-timesteps 100

3. Added --skip-final-eval flag for quick validation
   - Allows skipping time-consuming evaluation for quick tests
   - Useful for CPU-based architecture validation

VALIDATION RESULTS:
✅ mlp_baseline: Successfully trains 100 timesteps in ~20 seconds on CPU
❌ Graph architectures: Hang during training (performance issue identified)

GRAPH PERFORMANCE ISSUE:
All graph-based architectures (vision_free, full_hgt, gat, gcn, etc.) hang
during training on CPU. Investigation shows this is likely due to:
- Large graph dimensions (N_MAX_NODES=15856, NODE_FEATURE_DIM=55)
- Complex GNN operations (especially HGT) on CPU
- No GPU acceleration for graph processing

TASK DOCUMENT:
Created TASK_GRAPH_OPTIMIZATION.md with comprehensive investigation plan:
- Task 1: Investigate graph dimension reduction
- Task 2: Optimize graph processing performance
- Task 3: Add lightweight vision-free architectures (GAT, GCN variants)
- Task 4: Validate all architectures
- Includes debugging commands, references, and success criteria

TESTING:
# Working non-graph architecture:
python scripts/train_and_compare.py --experiment-name test \
    --architectures mlp_baseline --total-timesteps 100 --num-envs 2 \
    --skip-final-eval --train-dataset ../nclone/datasets/train \
    --test-dataset ../nclone/datasets/test --output-dir experiments/

Modified files:
- npp_rl/feature_extractors/configurable_extractor.py (lines 354, 369, 385)
- npp_rl/models/hgt_encoder.py (edge_index conversion)
- npp_rl/training/architecture_trainer.py (dynamic n_steps adjustment)
- scripts/train_and_compare.py (--skip-final-eval flag)
- TASK_GRAPH_OPTIMIZATION.md (new task document)

+++ npp_rl/feature_extractors/configurable_extractor.py
@@ -351,7 +351,7 @@ class ConfigurableMultimodalExtractor(BaseFeaturesExtractor):
                 # nclone environment already provides these keys directly
                 hgt_graph_obs = {
                     "graph_node_feats": observations["graph_node_feats"].float(),
-                    "graph_edge_index": observations["graph_edge_index"],
+                    "graph_edge_index": observations["graph_edge_index"].long(),  # Convert for indexing
                     "graph_edge_feats": observations["graph_edge_feats"].float(),
                     "graph_node_mask": observations["graph_node_mask"],
                     "graph_edge_mask": observations["graph_edge_mask"],
@@ -366,7 +366,7 @@ class ConfigurableMultimodalExtractor(BaseFeaturesExtractor):
             elif isinstance(self.graph_encoder, SimplifiedHGTEncoder):
                 # SimplifiedHGTEncoder takes separate arguments
                 node_features = observations["graph_node_feats"].float()
-                edge_index = observations["graph_edge_index"]
+                edge_index = observations["graph_edge_index"].long()  # Convert for indexing
                 node_mask = observations["graph_node_mask"]
                 node_types = observations.get(
                     "graph_node_types",
@@ -382,7 +382,7 @@ class ConfigurableMultimodalExtractor(BaseFeaturesExtractor):
             else:
                 # GAT and GCN take separate arguments (no type information)
                 node_features = observations["graph_node_feats"].float()
-                edge_index = observations["graph_edge_index"]
+                edge_index = observations["graph_edge_index"].long()  # Convert for indexing
                 node_mask = observations["graph_node_mask"]
                 _, graph_features = self.graph_encoder(
                     node_features, edge_index, node_mask
